Comprehensive Security Hardening and Finance Module Implementation (#13)
* chore: comprehensive security hardening, finance module implementation, and database refactoring


* fix: resolve CI failure and stabilize security/finance updates


---------

+++ innovation/app/models/finance.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime, Float, ForeignKey
+from sqlalchemy import Column, Integer, String, Text, DateTime, Numeric, ForeignKey
 from sqlalchemy.orm import relationship
 from datetime import datetime
 from ..db.database import Base
@@ -8,7 +8,7 @@ class Transaction(Base):
     
     id = Column(Integer, primary_key=True, index=True)
     description = Column(String(200), nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(10, 2), nullable=False)
     type = Column(String(20), nullable=False) # income, expense
     status = Column(String(20), default="pending") # pending, paid, cancelled
